Create Job decsription
For a dynamic company with a young team in the Netherlands we are looking for test engineers. Our client had impressive growth in the past 30 years resulting in employing over 1000 employees worldwide over 10 locations. In the Netherlands, they have 140 employees with nearly 50 different
nationalities. You obtain a healthy diet or you are thinking of starting one? Every week fresh fruit is delivered to the office and in your lunch breaks, you can challenge your colleagues in a game of foosball or table tennis.
Whether you commute by car or public transportation, the office is easy to reach as it is situated next to the highway and the train station.

Role description

This company has over 450 respectable clients in the gaming industry. Their clients are manufacturers of games which need to be tested according to international legislation and standards. You will be responsible for both iGaming testing and embedded testing. Every week new games need to be tested so you have a nice variety in your job. You will learn on the job until you have mastered the complete process resulting in all-round testing experience.

Role demands

• Bachelor or Master degree in Software Engineering, Computer Sciences or a comparable education
• Knowledge of Testing and working experience is an advantage
• Excellent knowledge of Java, C/ C++
• Knowledge of Python or SQL is an advantage
• You master the English language, spoken and written
• Valid working permit

Rewards

• Monthly salary of 2833
• Pension plan
• Travel allowance fully reimbursed
• Training
• 25 holidays
• International FUN working environment
• Directly employed by our client

Are you interested in delivering fair products of high quality with a diverse group of people from different nationalities? Contact Ioana Badea on +31 20 614 27 11 for more information.